chore: renamed transport file to bridge
