+++ packages/core/src/index.ts
@@ -4,4 +4,4 @@ export * from "./resources.js";
 export type * from "./types.js";
 export * from "./utils.js";
 export * from "./validator.js";
-export * from "./transport.js";
+export * from "./bridge.js";
